+++ src/components/options/tabs/tools-tab.ts
@@ -31,6 +31,9 @@ const TEMPLATE = /* html */ `
     <button id="editEmblemButton" data-tip="Click to open Emblem Editor" data-shortcut="Shift + Y">
       Emblems
     </button>
+    <button id="overviewFeaturesButton" data-tip="Click to open Geographical Features Overview" data-shortcut="Shift + F">
+      Features
+    </button>
     <button id="editGoods" data-tip="Click to open Goods Editor" data-shortcut="Shift + G">Goods</button>
     <button
       id="editHeightmapButton"